requirements.txt
aiocqhttp==1.3.0
aiofiles==0.6.0
aiogram==2.11.2
aiohttp==3.7.3
aiosqlite==0.16.0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
blinker==1.4
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coloredlogs==14.0
cssselect2==0.4.1
defusedxml==0.6.0
discord==1.0.1
discord.py==1.5.1
ffmpy==0.2.3
filetype==1.0.7
future==0.18.2
h11==0.11.0
h2==4.0.0
hpack==4.0.0
httpcore==0.12.2
httpx==0.16.1
humanfriendly==8.2
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
imageio==2.9.0
iso8601==0.1.13
itsdangerous==1.1.0
janus==0.6.1
Jinja2==2.11.2
line-bot-sdk==1.17.0
lottie==0.6.6
mando==0.6.4
MarkupSafe==1.1.1
multidict==5.0.2
numpy==1.19.4
Pillow==8.0.1
priority==1.3.0
pycparser==2.20
pydantic==1.7.2
PyPika==0.44.0
python-mirai-core==0.8.3
pytz==2020.4
PyYAML==5.3.1
Quart==0.13.1
radon==4.3.2
requests==2.25.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
tinycss2==1.1.0
toml==0.10.2
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
umr-coolq-driver==0.3.1
umr-discord-driver==0.2
umr-extensions-demo==0.2
umr-line-driver==0.2
umr-mirai-driver==0.5
umr-telegram-driver==0.2.1
unified-message-relay==4.2
urllib3==1.26.2
Wand==0.6.4
webencodings==0.5.1
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.6.3
